refactor: move git code to internal package
